project/backend/requirements.txt
Flask==1.1.1
requests==2.25.1
PyJWT==2.0.0
get-docker-secret==1.0.1
psycopg2-binary==2.8.5
Flask-Cors==3.0.10
